requirements.txt
Django==1.4.3
amqp==1.0.6
anyjson==0.3.3
billiard==2.7.3.19
celery==3.0.13
distribute==0.6.31
dj-database-url==0.2.1
django-celery==3.0.11
django-social-auth==0.7.18
gunicorn==0.17.2
httplib2==0.7.7
kombu==2.5.4
oauth2==1.5.211
psycopg2==2.4.5
python-dateutil==1.5
python-openid==2.2.5
pytz==2012j
wsgiref==0.1.2
